refactor(folder-structure): move databases to `db` folder
